payment-gateway-notifications/requirements.txt
# =============================================================================
# Payment Gateway Notifications - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
aiohttp>=3.9.0              # Async HTTP client/server for API and webhooks
python-dotenv>=1.0.0        # Environment variable management

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
asyncpg>=0.29.0             # PostgreSQL async driver
aiosqlite>=0.19.0           # SQLite async driver (for development)

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
pydantic>=2.5.0             # Data validation (optional, for stricter models)

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=7.4.0               # Testing framework
pytest-asyncio>=0.21.0      # Async test support
pytest-cov>=4.1.0           # Coverage reporting
aioresponses>=0.7.0         # Mock aiohttp requests

# -----------------------------------------------------------------------------
# Development
# -----------------------------------------------------------------------------
black>=23.0.0               # Code formatting
isort>=5.12.0               # Import sorting
flake8>=6.1.0               # Linting
mypy>=1.7.0                 # Type checking

# -----------------------------------------------------------------------------
# Production (optional)
# -----------------------------------------------------------------------------
uvloop>=0.19.0;platform_system!='Windows'  # Fast event loop for Linux/Mac
gunicorn>=21.0.0            # Production WSGI server (for additional processes)
